Fix two small visual bugs
When using this, I found that the item lines in the inventory where displayed with an unnecessary margin on the h4 title resulting in a broken display.
Also displaying the delete icon as `block` makes it appear an a new line, breaking the inventory item line as well. Displaying it as `inline` makes it appear beside its siblings.