fix checkDeathSaveStatus html search
checkDeathSaveStatus was attempting to find the sheet's html using a raw jquery `$` rather than html.find, meaning that it would fail when the rendered html was not in the same window (such as when the character sheet is popped out). This fixes the "dead" class not being applied when HP is reduced to zero on a popped-out character sheet window.